+++ processor/tunachain_processor/speciess/state.py
@@ -1,4 +1,4 @@
-# Speciess State
+# Species State
 #
 # Written by Mohammed Alsadi
 # -----------------------------------------------------------------------------
@@ -11,19 +11,19 @@ import logging
 LOGGER = logging.getLogger(__name__)
 
 
-SPECIESS_NAMESPACE = hashlib.sha512(
-    'speciess'.encode('utf-8')).hexdigest()[0:6]
-LOGGER.info(SPECIESS_NAMESPACE)    
+SPECIES_NAMESPACE = hashlib.sha512(
+    'species'.encode('utf-8')).hexdigest()[0:6]
+LOGGER.info(SPECIES_NAMESPACE)    
 
 
-def _get_address(speciessId):
-    adr = hashlib.sha512(speciessId.encode('utf-8')).hexdigest()[:62]
+def _get_address(speciesId):
+    adr = hashlib.sha512(speciesId.encode('utf-8')).hexdigest()[:62]
     LOGGER.info(adr)
     return adr
 
 
-def _get_speciess_address(speciessId):
-    add =  SPECIESS_NAMESPACE + '00' + _get_address(speciessId)
+def _get_species_address(speciesId):
+    add =  SPECIES_NAMESPACE + '00' + _get_address(speciesId)
     LOGGER.info(add)
     return add
 
@@ -36,23 +36,23 @@ def _serialize(data):
     return json.dumps(data, sort_keys=True).encode('utf-8')
 
 
-class SpeciessState(object):
+class SpeciesState(object):
 
     TIMEOUT = 3
 
     def __init__(self, context):
         self._context = context
 
-    def get_speciess(self, speciessId):
-        return self._get_state(_get_speciess_address(speciessId))
+    def get_species(self, speciesId):
+        return self._get_state(_get_species_address(speciesId))
 
     
-    def set_speciess(self, speciessId, quantity, species, packageNum, launchDateTime):
-        address = _get_speciess_address(speciessId)
-        LOGGER.info('set_speciess method')
+    def set_species(self, speciesId, quantity, species, packageNum, launchDateTime):
+        address = _get_species_address(speciesId)
+        LOGGER.info('set_species method')
         LOGGER.info(address)
         state_data = _serialize(
-            {   "speciessId": speciessId,
+            {   "speciesId": speciesId,
                 "quantity": quantity,
                 "species": species,
                 "packageNum": packageNum,
