+++ server/state.js
@@ -12,7 +12,6 @@ const KEY_NAME = 'transfer-chain.keys'
 const API_URL = 'http://localhost:8000/api'
 
 const transactions=[
-    {family:"speciess",version:"0.1", prefix:""},
     {family:"company",version:"0.1", prefix:""},
     {family:"haul",vesion:"0.1", prefix:""},
     {family:"pallet",version:"0.1", prefix:""},
